+++ integration_test/app_test.dart
@@ -18,7 +18,8 @@ void main() async {
     // First, define the Finders and use them to locate widgets from the
     // test suite. Note: the Strings provided to the `byValueKey` method must
     // be the same as the Strings we used for the Keys.
-    final loginButtonFinder = find.byKey(const Key('LoginPage_LoginButton'));
+    final loginButtonFinder =
+        find.byKey(const Key('LoginPage_LoginButton_Github'));
     final startListeningButtonFinder =
         find.byKey(const Key('LocationPage_StartListeningButton'));
     final locationTileFinder =